requirements.txt
fastapi
uvicorn
tensorflow
numpy
Pillow
opencv-python
